move parseOptions into separate module
