+++ lib/container/OverridablesPlugin.js
@@ -11,32 +11,13 @@ const OverridableModule = require("./OverridableModule");
 const OverridableOriginalDependency = require("./OverridableOriginalDependency");
 const OverridableOriginalModuleFactory = require("./OverridableOriginalModuleFactory");
 const OverridablesRuntimeModule = require("./OverridablesRuntimeModule");
+const parseOptions = require("./parseOptions");
 
 /** @typedef {import("../Compiler")} Compiler */
 
-const parseOptions = (prefix, options, result) => {
-	if (Array.isArray(options)) {
-		for (const item of options) {
-			parseOptions(prefix, item, result);
-		}
-	} else if (typeof options === "string") {
-		result.push([prefix + options.replace(/^([^\w]+\/)+/, ""), options]);
-	} else if (options && typeof options === "object") {
-		for (const key of Object.keys(options)) {
-			const value = options[key];
-			if (typeof value === "string") {
-				result.push([prefix + key, value]);
-			} else {
-				parseOptions(prefix + key + "/", value, result);
-			}
-		}
-	}
-};
-
 class OverridablesPlugin {
 	constructor(options) {
-		this.overridables = [];
-		parseOptions("", options, this.overridables);
+		this.overridables = parseOptions(options);
 	}
 	/**
 	 * @param {Compiler} compiler webpack compiler
@@ -95,6 +76,7 @@ class OverridablesPlugin {
 					"OverridablesPlugin",
 					(chunk, set) => {
 						set.add(RuntimeGlobals.module);
+						set.add(RuntimeGlobals.moduleFactories);
 						set.add(RuntimeGlobals.hasOwnProperty);
 						set.add(RuntimeGlobals.ensureChunkHandlers);
 						compilation.addRuntimeModule(
